+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 
 ### Improvements
 
+- Remove the Print settings in the Settings window (Instead, set print settings directly in the Print dialog).
 - [beta] Adjust layout in Settings window.